+++ src/components/card/SmallPetCard.tsx
@@ -1,8 +1,9 @@
+import { UtcStringToYearMonth } from "@/utils/dateConverter";
 import { Icon } from "@iconify/react";
 import { useMemo } from "react";
 import { Link } from "react-router-dom";
 import BlankPic144x144 from "../../assets/small-petcard/BlankPic144x144.svg";
-interface ISmallPetCardProps {
+interface SmallPetCardProps {
   id: string;
   name: string;
   species: string;
@@ -10,25 +11,6 @@ interface ISmallPetCardProps {
   birthdate: string;
   status: string;
 }
-const UtcStringToYearMonth = ({ utcString }: { utcString: string }) => {
-  const birthdate = new Date(utcString);
-  const currentUtcTime = new Date();
-
-  const timeDifference = currentUtcTime.getTime() - birthdate.getTime();
-  const yearsDifference = Math.floor(
-    timeDifference / (365 * 24 * 60 * 60 * 1000)
-  );
-  const remainingMonths = Math.floor(
-    (timeDifference % (365 * 24 * 60 * 60 * 1000)) /
-      (24 * 60 * 60 * 1000) /
-      (365 / 12)
-  );
-
-  const years = yearsDifference;
-  const months = remainingMonths;
-
-  return { years, months };
-};
 const smallpetcard = ({
   id,
   name,
@@ -36,9 +18,14 @@ const smallpetcard = ({
   gender,
   birthdate,
   status,
-}: ISmallPetCardProps) => {
-  const genderAns = gender == "male" ? "ผู้" : "เมีย";
-  const linkTo = "/pet/" + id;
+}: SmallPetCardProps) => {
+  const genderAns = useMemo(() => {
+    return gender === "male" ? "ผู้" : "เมีย";
+  }, [gender]);
+
+  const linkTo = useMemo(() => {
+    return "/pet/" + id;
+  }, [id]);
   const { years, months } = useMemo(() => {
     return UtcStringToYearMonth({ utcString: birthdate });
   }, [birthdate]);
@@ -46,12 +33,8 @@ const smallpetcard = ({
     <Link to={linkTo}>
       <div className="m-0 h-64 w-44 shrink-0 rounded-2xl bg-white p-3 shadow-md">
         <div className="flex flex-shrink-0 flex-col items-center justify-center gap-y-2.5">
-          <div className="bg-accent-grey flex w-36 rounded-2xl drop-shadow-[0_0_6px_rgba(0,0,0,0.50)]">
-            <img
-              className="rounded-3xl"
-              src={BlankPic144x144}
-              alt="PetPic 144*144"
-            />
+          <div className="flex w-36 rounded-2xl shadow-lg">
+            <img className="rounded-2xl" src={BlankPic144x144} alt={name} />
           </div>
 
           <div className="flex-start m-0 flex w-full flex-col items-start justify-center gap-2 p-0">
@@ -64,9 +47,7 @@ const smallpetcard = ({
                   <div>
                     <Icon
                       icon="ph:paw-print"
-                      color="#808086"
-                      width="20"
-                      height="20"
+                      className="h-5 w-5 text-accent-gray"
                     />
                   </div>
                   <div className="text-xs font-normal  text-accent-gray">
@@ -75,12 +56,7 @@ const smallpetcard = ({
                 </div>
                 <div className="flex-start m-0 flex flex-row items-center justify-start gap-2 p-0 ">
                   <div>
-                    <Icon
-                      icon="ph:gift"
-                      color="#808086"
-                      width="20"
-                      height="20"
-                    />
+                    <Icon icon="ph:gift" className="h-5 w-5 text-accent-gray" />
                   </div>
                   <div className="text-xs font-normal text-accent-gray">
                     อายุ {years} ปี {months} เดือน
@@ -93,12 +69,7 @@ const smallpetcard = ({
                 }`}
               >
                 <div className="px-1 py-1">
-                  <Icon
-                    icon="ph:check-bold"
-                    color="#FFFFFF"
-                    width="16"
-                    height="16"
-                  />
+                  <Icon icon="ph:check-bold" className="h-4 w-4 text-white" />
                 </div>
               </div>
             </div>
